fix(app): address review findings on the lifecycle lane
- Split TilingSettings' Codable half into
  TilingSettings+Coding.swift (379 lines breached the 350
  ceiling; both halves now in the sweet spot).
- SingleInstanceLock fails open on I/O errors: only a real
  flock contention (EWOULDBLOCK) reads as "already running" —
  an unopenable lock file must not block every future launch.
- QuitGridLayout pins cascaded frames inside the display so
  deep bottom-row piles (and minSize-floored cells on tiny
  displays) keep their title bars reachable; teardown is
  one-shot, nothing can correct an off-screen window after.
- Route quit.* in the outer dispatcher past layoutCommand's
  forced-retile trailer (track.swap precedent) — the setting
  is read only at quit, a live retile was a pure no-op.
- Derive quit.set_layout's error message from CaseIterable so
  a second style can't silently rot it.
- Add QuitCommandTests (dispatch seam), a fail-open lock test,
  a bottom-row pin test, and the single-window quarter-cell
  row in the accepted-limitations table.
